+++ _includes/includes/renderLanguageExample.php
@@ -1,36 +1,42 @@
 <?php /*
   Name:             renderLanguageExample
   Copyright:        Copyright (C) 2005 Tavultesoft Pty Ltd.
-  Documentation:    
-  Description:      
+  Documentation:
+  Description:
   Create Date:      21 Dec 2009
 
   Modified Date:    21 Dec 2009
   Authors:          mcdurdin
-  Related Files:    
-  Dependencies:     
+  Related Files:
+  Dependencies:
 
-  Bugs:             
-  Todo:             
-  Notes:            
+  Bugs:
+  Todo:
+  Notes:
   History:          21 Dec 2009 - mcdurdin - Fixup rendering of special keys
 */
   require_once('servervars.php');
-  
+
   function renderLanguageExample($keyboard, $language, $morehelp = '', $toBrowser = true)
   {
- 
+
     $keyboard = str_replace('Keyboard_','',$keyboard);
     $string = "https://api.keyman.com/cloud/3.0/keyboards/".$keyboard;
-    $json = file_get_contents($string);
-    $obj = json_decode($json,true);
+    if(($json = @file_get_contents($string)) === FALSE) {
+      return '';
+    }
+
+    if(($obj = @json_decode($json,true)) === NULL) {
+      return '';
+    }
+
     $result = '';
     $found = false;
 
     if($language == '') {
       // Show all examples
       $kbd = $obj['keyboard'];
-      
+
       foreach ($kbd['languages'] as $ExampleLanguage) {
         if(!isset($ExampleLanguage['example'])) {
           continue;
@@ -45,7 +51,7 @@
       	$result .= "</tr>";
       	$found = true;
       }
-      
+
       if($found) $result .= "</table>";
     }
     else if(isset($obj['keyboard']) && !empty($obj['keyboard']['languages'][0]['example'])) {
@@ -58,21 +64,21 @@
     if($toBrowser) echo $result;
     return $result;
   }
-  
+
   function DisplayExample($x, $table = false, $language = '')
   {
     global $morehelp;
     $result = '';
-    
+
     if(is_array($x))
     {
       $keys = KeyRenderer::renderInternal($x['keys'], $table);
       $out = $x['text'];
       if(isset($x['note'])) $note = " ({$x['note']})"; else $note = '';
-      
+
       if($table)
         $result .= "<td class='highlightExample lang2' lang='$language'>$out</td><td>{$x['note']}</td><td class='highlightKeys'>$keys</td>";
-      else        
+      else
       $result .=
         "To enter ".
         "<span class='highlightExample lang2' lang='$language'>$out</span>$note, ".
@@ -86,18 +92,18 @@
       $result .= "$x  $morehelp";
     return $result;
   }
-  
+
   class KeyRenderer
   {
     public static function render($s)
     {
       return "<span class='highlightKeys'>".KeyRenderer::renderInternal($s, true)."</span>";
     }
-    
+
     public static function renderInternal($s, $table = false)
     {
       $spkey = array("pgup","pgdn","left","right","up","down","ins","del","home","end","enter","tab","esc","caps","bksp","shift","ctrl","alt");
-      $spkey2 = array("shift","ctrl","alt");  
+      $spkey2 = array("shift","ctrl","alt");
 
       $i = 0;
       $r = '';
@@ -110,7 +116,7 @@
           $i++; $ch = $s[$i];
           if($ch == '[')
           {
-            $t = $ch;  
+            $t = $ch;
           }
           else if($ch != '[')
           {
@@ -127,19 +133,19 @@
 
             /* Special keys */
             $x = strpos($t1,'!'); $u = "";
-            if($x !== FALSE) 
+            if($x !== FALSE)
             {
               $ks = strtolower(substr($t1,$x+1,strlen($t1)-$x-1));
               if(array_search($ks,$spkey) !== FALSE)
-              { 
+              {
                 $u = $ks;
                 $kn = $ks;
                 if(array_search($ks,$spkey2) !== FALSE) $kn .= "2";
-                $u = "<img src='".cdn("img/keys/tr.gif")."' title='" . $ks ."' alt='" . $ks . "' class='key-" . $kn . "' />";                
+                $u = "<img src='".cdn("img/keys/tr.gif")."' title='" . $ks ."' alt='" . $ks . "' class='key-" . $kn . "' />";
                 $t = substr($t,0,$x); $ch = '';
               }
             }
-            
+
             if(strpos($t, 'C') !== FALSE) $r .= "<img src='".cdn("img/keys/tr.gif")."' title='Ctrl' alt='Ctrl' class='key-ctrl' />";
             if(strpos($t, 'S') !== FALSE) $r .= "<img src='".cdn("img/keys/tr.gif")."' title='Shift' alt='Shift' class='key-shift' />";
             if(strpos($t, 'A') !== FALSE) $r .= "<img src='".cdn("img/keys/tr.gif")."' title='Alt' alt='Alt' class='key-alt' />";
@@ -155,7 +161,7 @@
         if($endspan) $r .= "</span>";
         $i++;
       }
-      
+
       if($table)
       {
         $r .= "<span class='key-print'>";
@@ -183,13 +189,13 @@
 
               /* Special keys */
               $x = strpos($t1,'!'); $u = "";
-              if($x !== FALSE) 
+              if($x !== FALSE)
               {
                 $kn = substr($t1,$x+1,strlen($t1)-$x-1);
                 $ks = strtolower($kn);
                 if(array_search($ks,$spkey) !== FALSE)
-                { 
-                  $u = $kn;                
+                {
+                  $u = $kn;
                   $t = substr($t,0,$x); $ch = '';
                 }
               }
@@ -200,17 +206,17 @@
               $endspan = true;
             }
           }
-          
+
           if($ch == ' ') $r .= "<span class='key-grp'>Space</span>"; else if($ch <> '') $r .= htmlentities($ch, ENT_QUOTES);
           if($endspan) $r .= " </span>";
           $i++;
         }
         $r .= "</span>";
       }
-  
+
       return $r;
     }
   }
-  
+
   $keyRenderer = new KeyRenderer(); // Can also be used statically
 ?>