+++ developer/engine/web/14.0/reference/compatibility/ShowPinnedHelp.php
@@ -1,0 +1,25 @@
+<?php
+  require_once('includes/template.php');
+
+  head([
+    'title' => "Function: ShowPinnedHelp"
+  ]);
+
+  require_once('includes/__docheader.php');
+?>
+
+<h1 class="title" id="ShowPinnedHelp">ShowPinnedHelp</h1>
+
+<h2 id="Summary" name="Summary">Summary</h2>
+
+<p>Shows the on-screen keyboard and locks it at the resulting location.</p>
+
+<?php
+syntax("keyman.ShowPinnedHelp(x, y);");
+params();
+returns();
+?>
+
+<h3 id="ReplacedBy" name="ReplacedBy">Replaced by</h3>
+
+<p><a href="../osk/setRect"><code><?="keyman.osk.setRect()"?></code></a>, <a href="../osk/show"><code><?="keyman.osk.show()"?></code></a></p>