+++ developer/engine/web/history.php
@@ -17,7 +17,7 @@ head([
 
 require_once("includes/history_utils.php");
 
-display_history("web");
+display_history("web", "2.0");
 
 // Included to trick the side nav into generating a nice name for history.md in the sidebar.
 // <h1>KeymanWeb Version History</h1>