+++ products/desktop/14.0/docs/common_linux.php
@@ -1,8 +1,0 @@
-<?php
-  require_once('includes/template.php');
-
-  head([
-    'title' => "Does Keyman run on Linux?"
-  ]);
-?>
-<div class="navheader"><table width="100%" summary="Navigation header"><tr><th colspan="3" align="center">Does Keyman run on Linux?</th></tr><tr><td width="20%" align="left"><a accesskey="p" href="common_mac.php">Prev</a> </td><th width="60%" align="center">Chapter 5. Common Questions</th><td width="20%" align="right"> <a accesskey="n" href="common_osk.php">Next</a></td></tr></table><hr></div><div class="section" lang="en"><div class="titlepage"><div><div><h2 class="title" style="clear: both" id="common_linux">Does Keyman run on Linux?</h2></div></div></div><p>Yes. <a class="ulink" href="https://keyman.com/linux" target="_blank">Keyman for Linux</a>.</p></div><div class="navfooter"><hr><table width="100%" summary="Navigation footer"><tr><td width="40%" align="left"><a accesskey="p" href="common_mac.php">Prev</a> </td><td width="20%" align="center"><a accesskey="u" href="common.php">Up</a></td><td width="40%" align="right"> <a accesskey="n" href="common_osk.php">Next</a></td></tr><tr><td width="40%" align="left" valign="top">Does Keyman run on a Mac? </td><td width="20%" align="center"><a accesskey="h" href="index.php">Home</a></td><td width="40%" align="right" valign="top"> Does Keyman include a virtual on-screen keyboard?</td></tr></table></div>