+++ products/desktop/14.0/docs/common_old_version.php
@@ -1,9 +1,0 @@
-<?php
-  require_once('includes/template.php');
-
-  head([
-    'title' => "I am a registered Keyman 7.1 or Keyman Desktop 8 or Keyman Desktop 9 user. Can I still download Keyman 7.1 or Keyman Desktop 8 or Keyman Desktop 9?"
-  ]);
-?>
-<div class="navheader"><table width="100%" summary="Navigation header"><tr><th colspan="3" align="center">I am a registered Keyman 7.1 or Keyman Desktop 8 or Keyman Desktop 9 user. Can I still download Keyman 7.1 or Keyman Desktop 8 or Keyman Desktop 9?</th></tr><tr><td width="20%" align="left"><a accesskey="p" href="common_compatibility.php">Prev</a> </td><th width="60%" align="center">Chapter 5. Common Questions</th><td width="20%" align="right"> <a accesskey="n" href="common_support.php">Next</a></td></tr></table><hr></div><div class="section" lang="en"><div class="titlepage"><div><div><h2 class="title" style="clear: both" id="common_old_version">I am a registered Keyman 7.1 or Keyman Desktop 8 or Keyman Desktop 9 user. Can I still download Keyman 7.1 or Keyman Desktop 8 or Keyman Desktop 9?</h2></div></div></div><p>If you are a registered user of Keyman 7.1 or Keyman Desktop 8 or Keyman Desktop 9, you can download the most recent version these programs from <a class="ulink" href="http://keyman.com/downloads/archive/" target="_blank">keyman.com/downloads/archive/</a>.</p>&gt;
-</div><div class="navfooter"><hr><table width="100%" summary="Navigation footer"><tr><td width="40%" align="left"><a accesskey="p" href="common_compatibility.php">Prev</a> </td><td width="20%" align="center"><a accesskey="u" href="common.php">Up</a></td><td width="40%" align="right"> <a accesskey="n" href="common_support.php">Next</a></td></tr><tr><td width="40%" align="left" valign="top">Does Keyman work with all software? </td><td width="20%" align="center"><a accesskey="h" href="index.php">Home</a></td><td width="40%" align="right" valign="top"> What support do I get with my Keyman?</td></tr></table></div>