+++ products/linux/version-history/index.php
@@ -8,7 +8,7 @@ head([
 
 require_once("includes/history_utils.php");
 
-display_history("linux");
+display_history("linux", "2.0");
 
 // Included to trick the side nav into generating a nice name for history.md in the sidebar.
 // <h1>Keyman for Linux Version History</h1>